Drop the dependency on the syndication crate
The syndication crate doesn't add much value at the moment. When the
atom_syndication or rss crates released a semver-incompatible version,
syndication needs to be updated first. Also, when parsing fails, it
discards the errors and just returns a generic "Could not parse" error.
Let's handle the parsing ourselves so that we can recover and show the
error messages.